requirements.txt
discord.py-self
gspread
oauth2client
python-dotenv
colorama